+++ src/games/tank/src/objects/button/normalButton/buttonReplay.ts
@@ -2,7 +2,7 @@ import { Button } from "./button";
 
 export class ButtonReplay extends Button{
 	public handleOnPress(){
-        this.currentScene.events.emit('restartGame');
-        console.log("GameScene: ");
+		this.currentScene.events.emit('restartGame');
+		console.log("GameScene: ");
 	}
 }
