+++ src/games/tank/src/scenes/bootScene.ts
@@ -1,10 +1,6 @@
-import { AnimationHelper } from "../helpers/animation-helper";
-
 export class BootScene extends Phaser.Scene {
   private loadingBar: Phaser.GameObjects.Graphics;
   private progressBar: Phaser.GameObjects.Graphics;
-  // helpers
-  private animationHelperInstance: AnimationHelper;
 
   constructor() {
     super({
@@ -37,10 +33,6 @@ export class BootScene extends Phaser.Scene {
     this.load.on(
       'complete',
       function () {
-        this.animationHelperInstance = new AnimationHelper(
-          this,
-          this.cache.json.get('animationJSON')
-        );
         this.progressBar.destroy();
         this.loadingBar.destroy();
       },
