+++ src/games/tank/src/scenes/menuScene.ts
@@ -1,12 +1,7 @@
-import { ButtonStart } from "../objects/button/normalButton/buttonStart";
-import { ButtonMusic } from "../objects/button/toggleButton/buttonMusic";
-import { ButtonSound } from "../objects/button/toggleButton/buttonSound";
+import { MenuContainer } from "../objects/container/menuContainer";
 
 export class MenuScene extends Phaser.Scene {
-  private btnStart: ButtonStart;
-  private btnSound: ButtonSound;
-  private btnMusic: ButtonMusic;
-  private zone!: Phaser.GameObjects.Zone;
+  private menuContainer: MenuContainer; 
   private menuTrackAudio: Phaser.Sound.BaseSound;
   private menuStartAudio: Phaser.Sound.BaseSound;
   private switchScene!: boolean;
@@ -23,10 +18,9 @@ export class MenuScene extends Phaser.Scene {
   }
   
   create(): void {
-    
+    this.menuContainer = new MenuContainer(this, 0,0);
+    this.add.existing(this.menuContainer);
     this.initAudio();
-    this.createUI();
-    this.createTweens();
   }
 
   update(): void {
@@ -42,7 +36,6 @@ export class MenuScene extends Phaser.Scene {
         this.menuTrackAudio.pause();
       }
     }
-
   }
   
   private initAudio() {
@@ -52,90 +45,15 @@ export class MenuScene extends Phaser.Scene {
     this.menuStartAudio.on('complete', ()=>{
       this.menuStartAudio.pause();
       this.menuTrackAudio.play();
-      this.btnStart.setVisible(true);
+      this.menuContainer.setVisibleBtnStart();
     });
-  }
-
-  private createUI(){
-    this.add.image(0,0,'background')
-      .setOrigin(0,0)
-      .setScale(0.75,0.75);
-    
-    const logo = this.add.image(0,0,'logo');
-    this.zone = this.add.zone(140, 90, this.cameras.main.width - 140*2, this.cameras.main.height - 90*2).setOrigin(0,0)
-    //  Center the picture in the game
-    
-    this.btnStart = new ButtonStart({
-      scene: this,
-      x: 0,
-      y: 0,
-      texture: 'btn-start',
-      soundPress: 'select',
-    }).setVisible(false);
-    
-    
-    this.btnSound =  new ButtonSound({
-      scene: this,
-      x: 0,
-      y: 0,
-      texture: 'btn-sound',
-      frame: 1,
-      numberOfFrames: 2,
-      soundPress: 'click',
-    })
-    
-    this.btnMusic =  new ButtonMusic({
-      scene: this,
-      x: 0,
-      y: 0,
-      texture: 'btn-music',
-      frame: 1,
-      numberOfFrames: 2,
-      soundPress: 'click',
-    })
-    
-    Phaser.Display.Align.In.TopCenter(logo, this.zone);
-    Phaser.Display.Align.In.Center(this.btnStart, this.zone);
-    Phaser.Display.Align.In.BottomLeft(this.btnSound, this.zone);
-    Phaser.Display.Align.In.BottomRight(this.btnMusic, this.zone);
-
-    this.add.existing(this.btnSound);
-  }
-  
-  private createTweens(): void {
-    this.tweens.add({
-      targets: this.btnStart,
-      scaleX: 1.2,
-      scaleY: 1.2,
-      ease: 'Sine.easeInOut',
-      duration: 500,
-      yoyo: true,
-      repeat: -1,
-    })
   }
 
   private createHandleEvents(): void {
     this.events.on('startGame', ()=>{
       this.menuTrackAudio.stop();
       this.switchScene = true;
-      this.tweens.add({
-        targets: [this.btnSound, this.btnMusic],
-        y: this.cameras.main.height + 200,
-        ease: 'Power1',
-        duration: 500,
-      });
-      this.tweens.add({
-        targets: [this.btnStart],
-        y: -200,
-        ease: 'Power1',
-        duration: 500,
-        onComplete: () => {
-          this.scene.stop("MenuScene")
-          this.game.input.mouse.requestPointerLock();
-          this.scene.stop("GameScene");
-          this.scene.start("GameScene");
-        }
-      });
+      this.menuContainer.createTweenClose();
     }, this);
 
     this.events.on('start', ()=>{
